✅ Update plurals tests
